fixed warnings about integer conversion in DREAM

+++ DREAM/tasdreamExternalTests.hpp
@@ -84,7 +84,7 @@ private:
     int num_mc;
     int wfirst, wsecond, wthird;
     bool verbose;
-    long int rngseed;
+    int rngseed;
 
     TasDREAM::CppUniformSampler u;
